moved camera input scripts
