+++ Assets/Scripts/Tools/ScrollInputScalerToggle.cs
@@ -1,6 +1,7 @@
 using System;
 using System.Collections;
 using System.Collections.Generic;
+using Netherlands3D.Twin.Cameras.Input;
 using UnityEngine;
 using UnityEngine.EventSystems;
 using UnityEngine.UI;
