+++ Assets/_Application/Cameras/Input/BaseCameraInputProvider.cs.meta
