+++ Assets/_Functionalities/ObjectInformation/Scripts/BagInspector.cs
@@ -8,8 +8,8 @@ using Netherlands3D.GeoJSON;
 using Netherlands3D.SelectionTools;
 using Netherlands3D.SubObjects;
 using Netherlands3D.Twin;
+using Netherlands3D.Twin.Cameras.Input;
 using Netherlands3D.Twin.Layers;
-using Netherlands3D.Twin.Layers.LayerTypes;
 using Netherlands3D.Twin.Layers.LayerTypes.GeoJsonLayers;
 using TMPro;
 using UnityEngine;
